requirements.txt
fastapi
websockets
openai==0.28
firebase-admin
python-multipart
uvicorn[standard]
python-dotenv